+++ crates/biome_js_analyze/tests/specs/nursery/useSemanticElements/valid.jsx
@@ -1,0 +1,14 @@
+export const Component = () => (
+    <div>
+        hello world
+        <header>header</header>
+        <img alt="" src="image.jpg" ></img>
+    </div>
+);
+
+
+export const Component2 = () => (
+    <div aria-label="foo">
+        hello world
+    </div>
+);