+++ crates/biome_js_analyze/tests/specs/nursery/useSortedClasses/templateLiteralSpace.jsx
@@ -1,0 +1,4 @@
+<>
+	<div class={`${variable} px-2 foo p-4 bar`}/>
+	<div class={`px-2 foo p-4 bar ${variable}`}/>
+</>